register/email confirm improvement

+++ api/routes/auth/controllers/confirmEmail.js
@@ -7,12 +7,22 @@ module.exports = async (req, res) => {
         const confirmElement = await db.emailConfirm.findOne({ encryptedEmail: email }).select('userID')
         if (!confirmElement) return res.status(400).send({ success: false, message: 'DB error. Try again later or try create new account.' })
 
+        const identity = confirmElement.sensorID
+        const existSensor = await db.agroGsmSensors.findOne({ identity })
+        if (!existSensor) return res.status(400).send({ success: false, error: 'sensor with this identity does not exist' })
         // створємо фільтр пошуку та параметр редагування запису користувача
         const filter = { _id: confirmElement.userID }
         const update = { confirmed: true }
 
         // активація користувача
         const confirmedUser = await db.users.findOneAndUpdate(filter, update, { new: true })
+        confirmedUser.devices = [existSensor._id, ...confirmedUser.devices]
+
+        existSensor.user = confirmedUser._id
+        existSensor.name = identity
+
+        await confirmedUser.save()
+        await existSensor.save()
 
         // видаляємо документ активації так як користувача уже активовано
         const deleteDocument = await db.emailConfirm.findOneAndDelete({ encryptedEmail: email })
